Flatten nested module in xo_rust example
